1) register a single resource store for IdentityServer, 2) register the IS4 IPersistedGrantStore and not our internal one
